RESTService/requirements.txt
Flask==3.0.2
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
flask-swagger-ui==3.36.0
